Front END Question CRUD

+++ resources/assets/js/components/forum/ShowQuestion.vue
@@ -6,23 +6,25 @@
                     <div class="headline">
                         {{ data.title }}
                     </div>
-                    <span class="grey--text">{{ data.user }} said {{ data.created_at }}</span>
 
+                    <span class="grey--text">{{ data.user }} said {{ data.created_at }}</span>
                 </div>
+
                 <v-spacer></v-spacer>
-                <v-btn color="teal">5 Replies</v-btn>
+                <v-btn color="teal" dark>5 Replies</v-btn>
             </v-card-title>
 
             <v-card-text v-html="body"></v-card-text>
 
             <v-card-actions v-if="own">
-                <v-btn icon small>
+                <v-btn icon small @click="edit">
                     <v-icon color="orange">edit</v-icon>
                 </v-btn>
 
                 <v-btn icon small @click="destroy">
                     <v-icon color="red">delete</v-icon>
                 </v-btn>
+
             </v-card-actions>
         </v-container>
     </v-card>
@@ -30,10 +32,7 @@
 
 <script>
 
-
-
     export default {
-
         components: {},
         props: ['data'],
         data() {
@@ -48,10 +47,14 @@
         },
         methods: {
             destroy() {
-                axios.delete(`/api/question/${this.data.slug}`)
-                    .then(res => this.$router.push('/forum'))
-                    .catch(error => console.log(error.response.data))
+                if (confirm('Are you sure??'))
+                    axios.delete(`/api/question/${this.data.slug}`)
+                        .then(res => this.$router.push('/forum'))
+                        .catch(error => console.log(error.response.data))
 
+            },
+            edit() {
+                EventBus.$emit('startEditing')
             }
         }
     }
